pyproject.toml
[project]
name = "genai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "langgraph>=0.4.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.0",
]
